requirements.txt
# Python Requirements for Chat Project

bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
